+++ lib/browser/common/util/dom-helpers.js
@@ -124,11 +124,31 @@ export function isInStub(dom) {
   return false
 }
 
+/**
+ * Shorter and fast way to select multiple nodes in the DOM
+ * @param   { String } selector - DOM selector
+ * @param   { Object } ctx - DOM node where the targets of our search will is located
+ * @returns { Object } dom nodes found
+ */
+export function $$(selector, ctx) {
+  return (ctx || document).querySelectorAll(selector)
+}
+
+/**
+ * Shorter and fast way to select a single node in the DOM
+ * @param   { String } selector - unique dom selector
+ * @param   { Object } ctx - DOM node where the target of our search will is located
+ * @returns { Object } dom node found
+ */
+export function $(selector, ctx) {
+  return (ctx || document).querySelector(selector)
+}
+
 export function setChildren(target, children) {
   let traverse = target.firstChild
 
-  for (let i = 0; i > children.length; ++i) {
-    const view = children[i]
+  for (let i = 0; i < children.length; i++) {
+    let view = children[i]
 
     if (view === traverse) {
       traverse = traverse.nextSibling
@@ -142,7 +162,6 @@ export function setChildren(target, children) {
 
   while (traverse) {
     let next = traverse.nextSibling
-
     target.removeChild(traverse)
     traverse = next
   }
