+++ lib/browser/common/util/misc.js
@@ -1,4 +1,7 @@
-import { isWritable } from './check'
+import {
+  isArray,
+  isWritable
+} from './check'
 
 /**
  * Specialized function for looping an array-like collection with `each={}`
@@ -87,4 +90,58 @@ export function extend(src) {
     }
   }
   return src
+}
+
+/**
+ * Simple object prototypal inheritance
+ * @param   { Object } parent - parent object
+ * @returns { Object } child instance
+ */
+export function inherit(parent) {
+  function Child() {}
+  Child.prototype = parent
+  return new Child()
+}
+
+/**
+ * Set the property of an object for a given key. If something already
+ * exists there, then it becomes an array containing both the old and new value.
+ * @param { Object } obj - object on which to set the property
+ * @param { String } key - property name
+ * @param { Object } value - the value of the property to be set
+ * @param { Boolean } ensureArray - ensure that the property remains an array
+ */
+export function arrayishAdd(obj, key, value, ensureArray) {
+  var dest = obj[key]
+  var isArr = isArray(dest)
+
+  if (dest && dest === value) return
+
+  // if the key was never set, set it once
+  if (!dest && ensureArray) obj[key] = [value]
+  else if (!dest) obj[key] = value
+  // if it was an array and not yet set
+  else if (!isArr || isArr && !contains(dest, value)) {
+    if (isArr) dest.push(value)
+    else obj[key] = [dest, value]
+  }
+}
+
+/**
+ * Removes an item from an object at a given key. If the key points to an array,
+ * then the item is just removed from the array.
+ * @param { Object } obj - object on which to remove the property
+ * @param { String } key - property name
+ * @param { Object } value - the value of the property to be removed
+ * @param { Boolean } ensureArray - ensure that the property remains an array
+*/
+export function arrayishRemove(obj, key, value, ensureArray) {
+  if (isArray(obj[key])) {
+    each(obj[key], function(item, i) {
+      if (item === value) obj[key].splice(i, 1)
+    })
+    if (!obj[key].length) delete obj[key]
+    else if (obj[key].length == 1 && !ensureArray) obj[key] = obj[key][0]
+  } else
+    delete obj[key] // otherwise just delete the key
 }
