+++ lib/browser/common/util/tag-helpers.js
@@ -3,6 +3,8 @@ import {
 } from './dom-helpers'
 
 import {
+  arrayishAdd,
+  arrayishRemove,
   contains,
   each
 } from './misc'
@@ -143,80 +145,6 @@ export function cleanUpData(data) {
 }
 
 /**
- * Shorter and fast way to select multiple nodes in the DOM
- * @param   { String } selector - DOM selector
- * @param   { Object } ctx - DOM node where the targets of our search will is located
- * @returns { Object } dom nodes found
- */
-export function $$(selector, ctx) {
-  return (ctx || document).querySelectorAll(selector)
-}
-
-/**
- * Shorter and fast way to select a single node in the DOM
- * @param   { String } selector - unique dom selector
- * @param   { Object } ctx - DOM node where the target of our search will is located
- * @returns { Object } dom node found
- */
-export function $(selector, ctx) {
-  return (ctx || document).querySelector(selector)
-}
-
-/**
- * Simple object prototypal inheritance
- * @param   { Object } parent - parent object
- * @returns { Object } child instance
- */
-export function inherit(parent) {
-  function Child() {}
-  Child.prototype = parent
-  return new Child()
-}
-
-/**
- * Set the property of an object for a given key. If something already
- * exists there, then it becomes an array containing both the old and new value.
- * @param { Object } obj - object on which to set the property
- * @param { String } key - property name
- * @param { Object } value - the value of the property to be set
- * @param { Boolean } ensureArray - ensure that the property remains an array
- */
-export function arrayishAdd(obj, key, value, ensureArray) {
-  var dest = obj[key]
-  var isArr = isArray(dest)
-
-  if (dest && dest === value) return
-
-  // if the key was never set, set it once
-  if (!dest && ensureArray) obj[key] = [value]
-  else if (!dest) obj[key] = value
-  // if it was an array and not yet set
-  else if (!isArr || isArr && !contains(dest, value)) {
-    if (isArr) dest.push(value)
-    else obj[key] = [dest, value]
-  }
-}
-
-/**
- * Removes an item from an object at a given key. If the key points to an array,
- * then the item is just removed from the array.
- * @param { Object } obj - object on which to remove the property
- * @param { String } key - property name
- * @param { Object } value - the value of the property to be removed
- * @param { Boolean } ensureArray - ensure that the property remains an array
-*/
-export function arrayishRemove(obj, key, value, ensureArray) {
-  if (isArray(obj[key])) {
-    each(obj[key], function(item, i) {
-      if (item === value) obj[key].splice(i, 1)
-    })
-    if (!obj[key].length) delete obj[key]
-    else if (obj[key].length == 1 && !ensureArray) obj[key] = obj[key][0]
-  } else
-    delete obj[key] // otherwise just delete the key
-}
-
-/**
  * Mount a tag creating new Tag instance
  * @param   { Object } root - dom node where the tag will be mounted
  * @param   { String } tagName - name of the riot tag we want to mount
@@ -244,6 +172,26 @@ export function mountTo(root, tagName, opts) {
   return tag
 }
 
+/**
+ * Unmount redundant tags
+ * @param { Array } items - array containing the current items to loop
+ * @param { Array } tags - array containing all the children tags
+ * @param { String } tagName - key used to identify the type of tag
+ * @param { Object } parent - parent tag to remove the child from
+ */
+export function unmountRedundant(items, tags, tagName, parent) {
+  let
+    i = tags.length,
+    j = items.length,
+    t
+
+  while (i > j) {
+    t = tags[--i]
+    tags.splice(i, 1)
+    t.unmount()
+    arrayishRemove(parent.tags, tagName, t, true)
+  }
+}
 
 /**
  * Adds the elements for a virtual tag
@@ -306,4 +254,41 @@ export function selectTags(tags) {
       let name = t.trim().toLowerCase()
       return list + `,[${RIOT_TAG_IS}="${name}"]`
     }, '')
+}
+
+/**
+ * Convert the item looped into an object used to extend the child tag properties
+ * @param   { Object } exprObj - object containing the keys used to extend the children tags
+ * @param   { * } key - value to assign to the new object returned
+ * @param   { * } pos - value containing the position of the item in the array
+ * @param   { Object } base - prototype object for the new item
+ * @returns { Object } - new object containing the values of the original item
+ *
+ * The variables 'key' and 'pos' are arbitrary.
+ * They depend on the collection type looped (Array, Object)
+ * and on the expression used on the each tag
+ *
+ */
+export function mkitem(exprObj, key, pos, base) {
+  const item = base ? Object.create(base) : {}
+  item[exprObj.key] = key
+  item[exprObj.pos] = pos
+  return item
+}
+
+/**
+ * Move the nested custom tags in non custom loop tags
+ * @param   { Object } child - non custom loop tag
+ * @param   { Number } i - current position of the loop tag
+ */
+export function moveNestedTags(child, i) {
+  Object.keys(child.tags).forEach(function(tagName) {
+    var tag = child.tags[tagName]
+    if (isArray(tag))
+      each(tag, function (t) {
+        moveChildTag(t, tagName, i)
+      })
+    else
+      moveChildTag(tag, tagName, i)
+  })
 }
