+++ lib/browser/tag/each.js
@@ -21,69 +21,14 @@ import {
   moveChildTag,
   each,
   makeVirtual,
-  moveVirtual
+  moveVirtual,
+  mkitem,
+  moveNestedTags,
+  unmountRedundant
 } from './../common/util'
 
 import { tmpl } from 'riot-tmpl'
 import Tag from './tag'
-
-/**
- * Convert the item looped into an object used to extend the child tag properties
- * @param   { Object } expr - object containing the keys used to extend the children tags
- * @param   { * } key - value to assign to the new object returned
- * @param   { * } val - value containing the position of the item in the array
- * @param   { Object } base - prototype object for the new item
- * @returns { Object } - new object containing the values of the original item
- *
- * The variables 'key' and 'val' are arbitrary.
- * They depend on the collection type looped (Array, Object)
- * and on the expression used on the each tag
- *
- */
-export function mkitem(expr, key, val, base) {
-  var item = base ? Object.create(base) : {}
-  item[expr.key] = key
-  if (expr.pos) item[expr.pos] = val
-  return item
-}
-
-/**
- * Unmount the redundant tags
- * @param   { Array } items - array containing the current items to loop
- * @param   { Array } tags - array containing all the children tags
- * @param   { String } tagName - key used to identify the type of tag
- * @param   { Object } parent - parent tag to remove the child from
- */
-export function unmountRedundant(items, tags, tagName, parent) {
-
-  var i = tags.length,
-    j = items.length,
-    t
-
-  while (i > j) {
-    t = tags[--i]
-    tags.splice(i, 1)
-    t.unmount()
-    arrayishRemove(parent.tags, tagName, t, true)
-  }
-}
-
-/**
- * Move the nested custom tags in non custom loop tags
- * @param   { Object } child - non custom loop tag
- * @param   { Number } i - current position of the loop tag
- */
-export function moveNestedTags(child, i) {
-  Object.keys(child.tags).forEach(function(tagName) {
-    var tag = child.tags[tagName]
-    if (isArray(tag))
-      each(tag, function (t) {
-        moveChildTag(t, tagName, i)
-      })
-    else
-      moveChildTag(tag, tagName, i)
-  })
-}
 
 /**
  * Manage tags having the 'each'
@@ -131,7 +76,6 @@ export default function _each(dom, parent, expr) {
 
     // object loop. any changes cause full redraw
     if (!isArray(items)) {
-      console.log(items)
       hasKeys = items || false
       items = hasKeys ?
         Object.keys(items).map(function (key) {
